+++ src/main/java/org/kumoricon/site/report/export/ExportPresenter.java
@@ -2,6 +2,7 @@ package org.kumoricon.site.report.export;
 
 import com.vaadin.server.StreamResource;
 import org.kumoricon.site.report.export.data.AttendeeTSVExport;
+import org.kumoricon.site.report.export.data.TillTSVExport;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,6 +14,9 @@ public class ExportPresenter {
     @Autowired
     private AttendeeTSVExport attendeeTSVExport;
 
+    @Autowired
+    private TillTSVExport tillTSVExport;
+
     private static final Logger log = LoggerFactory.getLogger(ExportPresenter.class);
 
 
@@ -23,5 +27,10 @@ public class ExportPresenter {
     public StreamResource createAttendeeExport(ExportView view) {
         log.info("{} downloaded Attendee Export", view.getCurrentUser());
         return attendeeTSVExport.getStream();
+    }
+
+    public StreamResource createTillExport(ExportView view) {
+        log.info("{} downloaded Till Export", view.getCurrentUser());
+        return tillTSVExport.getStream();
     }
 }